Original Pure Maze Gen
This is a maze generation done from this link:https://catlikecoding.com/unity/tutorials/maze/

We still need to find some way to implement the rest of the tutorial : making rooms and not just maze corridors, however we cant use doors and instead use empty walls.

If you want, try to do it, but ill probably get it done tommorow.